feat: Start with the `ChainOfCommand` class
This class will be used to determine specific values on a one-by-one
basis. For instance; to determine what configuration file will be used,
a series of checks has to be performed (first, check if the command line
argument is set, then the environment variable, then if a specific file
exists, etc.). The `ChainOfCommand` class can be used for this.