refactor: Start with a lot of refactoring.
The `Slough` and `SloughConfig` classes are now refactored for a simpler
and better code structure. To make sure they work as intended, I had to
update the tests too. This will only be for the good, ofcourse.

Still some TODO's in the code. They will be picked up later. The CLI
part of the application is also heavily broken since there are a lot of
methods missing or changed.

Still a work in progress.